Added doxygen comment template.
Added latex directory to .gitignore so that it is not always pushed. We should only push the final pdf when complete.
Added a doxygen comment template to command_handler.c, commands.c, and time.c. We will need to fill in this templates
and add a comment templete for .h files.